Fixed Specialisations between CS and IS
